Terraform/Azure/applications/function-apps1/requirements.txt
azure-functions
rapidfuzz==3.6.1